+++ sound/soc/codecs/wmfw.h
@@ -30,6 +30,7 @@
 #define WMFW_CTL_TYPE_ACKED       0x1000 /* acked control */
 #define WMFW_CTL_TYPE_HOSTEVENT   0x1001 /* event control */
 #define WMFW_CTL_TYPE_HOST_BUFFER 0x1002 /* host buffer pointer */
+#define WMFW_CTL_TYPE_FWEVENT     0x1004 /* firmware event control */
 
 struct wmfw_header {
 	char magic[4];
@@ -55,6 +56,10 @@ struct wmfw_adsp2_sizes {
 	__le32 ym;
 	__le32 pm;
 	__le32 zm;
+} __packed;
+
+struct wmfw_vpu_sizes {
+	__le32 dm;
 } __packed;
 
 struct wmfw_region {
@@ -88,6 +93,38 @@ struct wmfw_adsp2_id_hdr {
 	__be32 n_algs;
 } __packed;
 
+struct wmfw_vpu_fwid_hdr {
+	__be32 core_id;
+	__be32 block_rev;
+	__be32 vendor_id;
+	__be32 firmware_id;
+	__be32 ver;
+} __packed;
+
+struct wmfw_vpu_id_hdr {
+	struct wmfw_vpu_fwid_hdr fw;
+	__be32 dm_base;
+	__be32 dm_size;
+	__be32 n_algs;
+} __packed;
+
+struct wmfw_halo_fwid_hdr {
+	__be32 core_id;
+	__be32 block_rev;
+	__be32 vendor_id;
+	__be32 id;
+	__be32 ver;
+} __packed;
+
+struct wmfw_halo_id_hdr {
+	struct wmfw_halo_fwid_hdr fw;
+	__be32 xm_base;
+	__be32 xm_size;
+	__be32 ym_base;
+	__be32 ym_size;
+	__be32 n_algs;
+} __packed;
+
 struct wmfw_alg_hdr {
 	__be32 id;
 	__be32 ver;
@@ -104,6 +141,20 @@ struct wmfw_adsp2_alg_hdr {
 	__be32 zm;
 	__be32 xm;
 	__be32 ym;
+} __packed;
+
+struct wmfw_vpu_alg_hdr {
+	struct wmfw_alg_hdr alg;
+	__be32 dm_base;
+	__be32 dm_size;
+} __packed;
+
+struct wmfw_halo_alg_hdr {
+	struct wmfw_alg_hdr alg;
+	__be32 xm_base;
+	__be32 xm_size;
+	__be32 ym_base;
+	__be32 ym_size;
 } __packed;
 
 struct wmfw_adsp_alg_data {
@@ -154,9 +205,12 @@ struct wmfw_coeff_item {
 
 #define WMFW_ADSP1 1
 #define WMFW_ADSP2 2
+#define WMFW_HALO 4
+#define WMFW_VPU 0x45
 
 #define WMFW_ABSOLUTE         0xf0
 #define WMFW_ALGORITHM_DATA   0xf2
+#define WMFW_METADATA         0xfc
 #define WMFW_NAME_TEXT        0xfe
 #define WMFW_INFO_TEXT        0xff
 
@@ -168,5 +222,11 @@ struct wmfw_coeff_item {
 #define WMFW_ADSP2_ZM 4
 #define WMFW_ADSP2_XM 5
 #define WMFW_ADSP2_YM 6
+
+#define WMFW_HALO_PM_PACKED 0x10
+#define WMFW_HALO_XM_PACKED 0x11
+#define WMFW_HALO_YM_PACKED 0x12
+
+#define WMFW_VPU_DM 0x30
 
 #endif
