--- techpack/audio/asoc/codecs/Android.mk
@@ -1,167 +1,0 @@
-# Android makefile for audio kernel modules
-
-# Assume no targets will be supported
-
-# Check if this driver needs be built for current target
-ifeq ($(call is-board-platform-in-list,msmnile sdmshrike),true)
-ifeq ($(TARGET_BOARD_AUTO),true)
-AUDIO_SELECT  := CONFIG_SND_SOC_SA8155=m
-else
-AUDIO_SELECT  := CONFIG_SND_SOC_SM8150=m
-endif
-endif
-
-ifeq ($(call is-board-platform,$(MSMSTEPPE)),true)
-ifeq ($(TARGET_BOARD_AUTO),true)
-AUDIO_SELECT  := CONFIG_SND_SOC_SA6155=m
-else
-AUDIO_SELECT  := CONFIG_SND_SOC_SM6150=m
-endif
-endif
-
-ifeq ($(call is-board-platform,$(TRINKET)),true)
-AUDIO_SELECT  := CONFIG_SND_SOC_SM6150=m
-endif
-
-ifeq ($(call is-board-platform,kona),true)
-AUDIO_SELECT  := CONFIG_SND_SOC_KONA=m
-endif
-
-ifeq ($(call is-board-platform,lito),true)
-AUDIO_SELECT  := CONFIG_SND_SOC_LITO=m
-endif
-
-ifeq ($(call is-board-platform,bengal),true)
-AUDIO_SELECT  := CONFIG_SND_SOC_BENGAL=m
-endif
-
-ifeq ($(call is-board-platform,sdm660),true)
-AUDIO_SELECT  := CONFIG_SND_SOC_SDM660=m
-endif
-
-AUDIO_CHIPSET := audio
-# Build/Package only in case of supported target
-ifeq ($(call is-board-platform-in-list,msmnile $(MSMSTEPPE) $(TRINKET) kona lito bengal sdmshrike sdm660),true)
-
-LOCAL_PATH := $(call my-dir)
-
-# This makefile is only for DLKM
-ifneq ($(findstring vendor,$(LOCAL_PATH)),)
-
-ifneq ($(findstring opensource,$(LOCAL_PATH)),)
-	AUDIO_BLD_DIR := $(shell pwd)/vendor/qcom/opensource/audio-kernel
-endif # opensource
-
-DLKM_DIR := $(TOP)/device/qcom/common/dlkm
-
-# Build audio.ko as $(AUDIO_CHIPSET)_audio.ko
-###########################################################
-# This is set once per LOCAL_PATH, not per (kernel) module
-KBUILD_OPTIONS := AUDIO_ROOT=$(AUDIO_BLD_DIR)
-
-# We are actually building audio.ko here, as per the
-# requirement we are specifying <chipset>_audio.ko as LOCAL_MODULE.
-# This means we need to rename the module to <chipset>_audio.ko
-# after audio.ko is built.
-KBUILD_OPTIONS += MODNAME=wcd_core_dlkm
-KBUILD_OPTIONS += BOARD_PLATFORM=$(TARGET_BOARD_PLATFORM)
-KBUILD_OPTIONS += $(AUDIO_SELECT)
-# Target specific build
-###########################################################
-ifneq ($(TARGET_BOARD_AUTO),true)
-include $(CLEAR_VARS)
-LOCAL_MODULE              := $(AUDIO_CHIPSET)_wcd_core.ko
-LOCAL_MODULE_KBUILD_NAME  := wcd_core_dlkm.ko
-LOCAL_MODULE_TAGS         := optional
-LOCAL_MODULE_DEBUG_ENABLE := true
-LOCAL_MODULE_PATH         := $(KERNEL_MODULES_OUT)
-include $(DLKM_DIR)/AndroidKernelModule.mk
-###########################################################
-include $(CLEAR_VARS)
-LOCAL_MODULE              := $(AUDIO_CHIPSET)_wcd9xxx.ko
-LOCAL_MODULE_KBUILD_NAME  := wcd9xxx_dlkm.ko
-LOCAL_MODULE_TAGS         := optional
-LOCAL_MODULE_DEBUG_ENABLE := true
-LOCAL_MODULE_PATH         := $(KERNEL_MODULES_OUT)
-include $(DLKM_DIR)/AndroidKernelModule.mk
-###########################################################
-ifeq ($(call is-board-platform-in-list, ),true)
-include $(CLEAR_VARS)
-LOCAL_MODULE              := $(AUDIO_CHIPSET)_wcd_cpe.ko
-LOCAL_MODULE_KBUILD_NAME  := wcd_cpe_dlkm.ko
-LOCAL_MODULE_TAGS         := optional
-LOCAL_MODULE_DEBUG_ENABLE := true
-LOCAL_MODULE_PATH         := $(KERNEL_MODULES_OUT)
-include $(DLKM_DIR)/AndroidKernelModule.mk
-endif
-###########################################################
-ifeq ($(call is-board-platform-in-list,msmnile $(MSMSTEPPE) $(TRINKET) sdmshrike),true)
-include $(CLEAR_VARS)
-LOCAL_MODULE              := $(AUDIO_CHIPSET)_wcd_spi.ko
-LOCAL_MODULE_KBUILD_NAME  := wcd_spi_dlkm.ko
-LOCAL_MODULE_TAGS         := optional
-LOCAL_MODULE_DEBUG_ENABLE := true
-LOCAL_MODULE_PATH         := $(KERNEL_MODULES_OUT)
-include $(DLKM_DIR)/AndroidKernelModule.mk
-endif
-###########################################################
-ifeq ($(call is-board-platform-in-list, sdm660),true)
-include $(CLEAR_VARS)
-LOCAL_MODULE              := $(AUDIO_CHIPSET)_wcd9335.ko
-LOCAL_MODULE_KBUILD_NAME  := wcd9335_dlkm.ko
-LOCAL_MODULE_TAGS         := optional
-LOCAL_MODULE_DEBUG_ENABLE := true
-LOCAL_MODULE_PATH         := $(KERNEL_MODULES_OUT)
-include $(DLKM_DIR)/AndroidKernelModule.mk
-endif
-###########################################################
-ifneq ($(call is-board-platform-in-list, bengal),true)
-include $(CLEAR_VARS)
-LOCAL_MODULE              := $(AUDIO_CHIPSET)_wsa881x.ko
-LOCAL_MODULE_KBUILD_NAME  := wsa881x_dlkm.ko
-LOCAL_MODULE_TAGS         := optional
-LOCAL_MODULE_DEBUG_ENABLE := true
-LOCAL_MODULE_PATH         := $(KERNEL_MODULES_OUT)
-include $(DLKM_DIR)/AndroidKernelModule.mk
-endif
-###########################################################
-ifeq ($(call is-board-platform-in-list, bengal),true)
-include $(CLEAR_VARS)
-LOCAL_MODULE              := $(AUDIO_CHIPSET)_wsa881x_analog.ko
-LOCAL_MODULE_KBUILD_NAME  := wsa881x_analog_dlkm.ko
-LOCAL_MODULE_TAGS         := optional
-LOCAL_MODULE_DEBUG_ENABLE := true
-LOCAL_MODULE_PATH         := $(KERNEL_MODULES_OUT)
-include $(DLKM_DIR)/AndroidKernelModule.mk
-endif
-###########################################################
-include $(CLEAR_VARS)
-LOCAL_MODULE              := $(AUDIO_CHIPSET)_mbhc.ko
-LOCAL_MODULE_KBUILD_NAME  := mbhc_dlkm.ko
-LOCAL_MODULE_TAGS         := optional
-LOCAL_MODULE_DEBUG_ENABLE := true
-LOCAL_MODULE_PATH         := $(KERNEL_MODULES_OUT)
-include $(DLKM_DIR)/AndroidKernelModule.mk
-endif # target specific build
-###########################################################
-include $(CLEAR_VARS)
-LOCAL_MODULE              := $(AUDIO_CHIPSET)_stub.ko
-LOCAL_MODULE_KBUILD_NAME  := stub_dlkm.ko
-LOCAL_MODULE_TAGS         := optional
-LOCAL_MODULE_DEBUG_ENABLE := true
-LOCAL_MODULE_PATH         := $(KERNEL_MODULES_OUT)
-include $(DLKM_DIR)/AndroidKernelModule.mk
-###########################################################
-ifneq ($(call is-board-platform-in-list, bengal),true)
-include $(CLEAR_VARS)
-LOCAL_MODULE              := $(AUDIO_CHIPSET)_hdmi.ko
-LOCAL_MODULE_KBUILD_NAME  := hdmi_dlkm.ko
-LOCAL_MODULE_TAGS         := optional
-LOCAL_MODULE_DEBUG_ENABLE := true
-LOCAL_MODULE_PATH         := $(KERNEL_MODULES_OUT)
-include $(DLKM_DIR)/AndroidKernelModule.mk
-endif
-###########################################################
-
-endif # DLKM check
-endif # supported target check
