--- techpack/audio/asoc/codecs/aqt1000/Android.mk
@@ -1,41 +1,0 @@
-# Android makefile for audio kernel modules
-
-AUDIO_CHIPSET := audio
-# Build/Package only in case of supported target
-ifeq ($(call is-board-platform-in-list,sdm670),true)
-
-LOCAL_PATH := $(call my-dir)
-
-# This makefile is only for DLKM
-ifneq ($(findstring vendor,$(LOCAL_PATH)),)
-
-ifneq ($(findstring opensource,$(LOCAL_PATH)),)
-	AUDIO_BLD_DIR := $(shell pwd)/vendor/qcom/opensource/audio-kernel
-endif # opensource
-
-DLKM_DIR := $(TOP)/device/qcom/common/dlkm
-
-# Build audio.ko as $(AUDIO_CHIPSET)_audio.ko
-###########################################################
-# This is set once per LOCAL_PATH, not per (kernel) module
-KBUILD_OPTIONS := AUDIO_ROOT=$(AUDIO_BLD_DIR)
-
-# We are actually building audio.ko here, as per the
-# requirement we are specifying <chipset>_audio.ko as LOCAL_MODULE.
-# This means we need to rename the module to <chipset>_audio.ko
-# after audio.ko is built.
-KBUILD_OPTIONS += MODNAME=aqt1000_cdc_dlkm
-KBUILD_OPTIONS += BOARD_PLATFORM=$(TARGET_BOARD_PLATFORM)
-KBUILD_OPTIONS += $(AUDIO_SELECT)
-
-###########################################################
-include $(CLEAR_VARS)
-LOCAL_MODULE              := $(AUDIO_CHIPSET)_aqt1000_cdc.ko
-LOCAL_MODULE_KBUILD_NAME  := aqt1000_cdc_dlkm.ko
-LOCAL_MODULE_TAGS         := optional
-LOCAL_MODULE_DEBUG_ENABLE := true
-LOCAL_MODULE_PATH         := $(KERNEL_MODULES_OUT)
-include $(DLKM_DIR)/AndroidKernelModule.mk
-###########################################################
-endif # DLKM check
-endif # supported target check
