+++ techpack/audio/config/litoauto.conf
@@ -1,47 +1,44 @@
-CONFIG_PINCTRL_LPI=y
-CONFIG_AUDIO_EXT_CLK=y
-CONFIG_SND_SOC_WCD9XXX_V2=y
-CONFIG_SND_SOC_WCD_MBHC=y
-CONFIG_SND_SOC_WSA881X=y
-CONFIG_SND_SOC_WSA883X=y
-CONFIG_WCD9XXX_CODEC_CORE_V2=y
-CONFIG_MSM_CDC_PINCTRL=y
-CONFIG_MSM_QDSP6V2_CODECS=y
-CONFIG_MSM_ULTRASOUND=y
-CONFIG_MSM_QDSP6_APRV2_RPMSG=y
-CONFIG_MSM_ADSP_LOADER=y
-CONFIG_REGMAP_SWR=y
-CONFIG_MSM_QDSP6_SSR=y
-CONFIG_MSM_QDSP6_PDR=y
-CONFIG_MSM_QDSP6_NOTIFIER=y
-CONFIG_SND_SOC_MSM_HOSTLESS_PCM=y
-CONFIG_SND_SOC_MSM_QDSP6V2_INTF=y
-CONFIG_SOUNDWIRE=y
-CONFIG_SOUNDWIRE_MSTR_CTRL=y
-CONFIG_SND_SOC_QDSP6V2=y
-CONFIG_SND_SOC_WCD_MBHC_ADC=y
-CONFIG_SND_SOC_MSM_HDMI_CODEC_RX=y
-CONFIG_QTI_PP=y
-CONFIG_SND_HWDEP_ROUTING=y
-CONFIG_SND_SOC_MSM_STUB=y
-CONFIG_MSM_AVTIMER=y
-CONFIG_SND_SOC_BOLERO=y
-CONFIG_WSA_MACRO=y
-CONFIG_VA_MACRO=y
-CONFIG_RX_MACRO=y
-CONFIG_TX_MACRO=y
-CONFIG_DIGITAL_CDC_RSC_MGR=y
-CONFIG_SND_SOC_WCD_IRQ=y
-CONFIG_SND_SOC_WCD938X=y
-CONFIG_SND_SOC_WCD938X_SLAVE=y
-CONFIG_SND_SOC_WCD937X=y
-CONFIG_SND_SOC_WCD937X_SLAVE=y
-CONFIG_SND_SOC_LITO=y
-CONFIG_SND_EVENT=y
-CONFIG_MI2S_DISABLE=y
-CONFIG_TDM_DISABLE=y
-CONFIG_AUXPCM_DISABLE=y
-CONFIG_DIGITAL_CDC_RSC_MGR=y
-CONFIG_SND_SOC_CS35L41=y
-CONFIG_MSM_CSPL=y
-
+CONFIG_PINCTRL_LPI=m
+CONFIG_AUDIO_EXT_CLK=m
+CONFIG_SND_SOC_WCD9XXX_V2=m
+CONFIG_SND_SOC_WCD_MBHC=m
+CONFIG_SND_SOC_WSA881X=m
+CONFIG_SND_SOC_WSA883X=m
+CONFIG_WCD9XXX_CODEC_CORE_V2=m
+CONFIG_MSM_CDC_PINCTRL=m
+CONFIG_MSM_QDSP6V2_CODECS=m
+CONFIG_MSM_ULTRASOUND=m
+CONFIG_MSM_QDSP6_APRV2_RPMSG=m
+CONFIG_MSM_ADSP_LOADER=m
+CONFIG_REGMAP_SWR=m
+CONFIG_MSM_QDSP6_SSR=m
+CONFIG_MSM_QDSP6_PDR=m
+CONFIG_MSM_QDSP6_NOTIFIER=m
+CONFIG_SND_SOC_MSM_HOSTLESS_PCM=m
+CONFIG_SND_SOC_MSM_QDSP6V2_INTF=m
+CONFIG_SOUNDWIRE=m
+CONFIG_SOUNDWIRE_MSTR_CTRL=m
+CONFIG_SND_SOC_QDSP6V2=m
+CONFIG_SND_SOC_WCD_MBHC_ADC=m
+CONFIG_SND_SOC_MSM_HDMI_CODEC_RX=m
+CONFIG_QTI_PP=m
+CONFIG_SND_HWDEP_ROUTING=m
+CONFIG_SND_SOC_MSM_STUB=m
+CONFIG_MSM_AVTIMER=m
+CONFIG_SND_SOC_BOLERO=m
+CONFIG_WSA_MACRO=m
+CONFIG_VA_MACRO=m
+CONFIG_RX_MACRO=m
+CONFIG_TX_MACRO=m
+CONFIG_DIGITAL_CDC_RSC_MGR=m
+CONFIG_SND_SOC_WCD_IRQ=m
+CONFIG_SND_SOC_WCD938X=m
+CONFIG_SND_SOC_WCD938X_SLAVE=m
+CONFIG_SND_SOC_WCD937X=m
+CONFIG_SND_SOC_WCD937X_SLAVE=m
+CONFIG_SND_SOC_LITO=m
+CONFIG_SND_EVENT=m
+CONFIG_TDM_DISABLE=m
+CONFIG_AUXPCM_DISABLE=m
+CONFIG_SND_SOC_CS47l35=m
+CONFIG_SND_SOC_CS35l41=m
