+++ techpack/audio/config/litoautoconf.h
@@ -42,9 +42,7 @@
 #define CONFIG_SND_SOC_WCD937X_SLAVE 1
 #define CONFIG_SND_SOC_LITO 1
 #define CONFIG_SND_EVENT 1
-#define CONFIG_MI2S_DISABLE 1
+#define CONFIG_TDM_DISABLE 1
 #define CONFIG_AUXPCM_DISABLE 1
-#define CONFIG_DIGITAL_CDC_RSC_MGR 1
-#define CONFIG_SND_SOC_CS35L41 1
-#define CONFIG_MSM_CSPL 1
-
+#define CONFIG_SND_SOC_CS47l35 1
+#define CONFIG_SND_SOC_CS35l41 1
